requirements-brief-e.txt
# Brief E additions · operator merges into pyproject.toml
# or installs: uv pip install -r requirements-brief-e.txt

fastapi>=0.115
uvicorn[standard]>=0.32
pydantic>=2.9
pydantic-settings>=2.6
httpx>=0.27
pyyaml>=6.0
tiktoken>=0.8

openai>=1.54
azure-identity>=1.19
azure-search-documents>=11.6.0b6
azure-ai-documentintelligence>=1.0.0b4
azure-ai-contentsafety>=1.0.0
azure-cosmos>=4.7
azure-monitor-opentelemetry>=1.6
azure-ai-ml>=1.21

docling>=2.8
pypdf>=5.0
asyncpg>=0.30

langchain>=0.3
langchain-core>=0.3
langgraph>=0.2.50
semantic-kernel>=1.14

mcp>=1.0

llama-index-core>=0.12
llama-index-embeddings-azure-openai>=0.3
llama-index-embeddings-ollama>=0.4
llama-index-vector-stores-azureaisearch>=0.3

colpali-engine>=0.3
pdf2image>=1.17
Pillow>=10.0

presidio-analyzer>=2.2
presidio-anonymizer>=2.2
nemoguardrails>=0.10

langfuse>=2.50
opentelemetry-sdk>=1.28

ragas>=0.2
datasets>=3.0

reportlab>=4.2

pytest>=8.3
pytest-asyncio>=0.24
